test/be/configure Jest properly and move shared constants to shared TR package
