+++ apps/backend/package.json
@@ -21,57 +21,41 @@
     "lint:eslint": "eslint \"{src,test}/**/*.{ts,js}\"",
     "lint:eslint:fix": "yarn lint:eslint --fix",
     "lint": "yarn lint:types && yarn lint:eslint",
-    "lint:fix": "yarn lint:eslint:fix"
+    "lint:fix": "yarn lint:eslint:fix",
+    "nst:upgrade": "npx ncu \"/nestjs*/\""
   },
   "dependencies": {
     "@apollo/server": "^4.9.3",
+    "@inno/constants": "*",
     "@inno/utils": "*",
     "@nestjs/apollo": "^12.0.9",
-    "@nestjs/common": "^9.0.0",
-    "@nestjs/core": "^9.0.0",
+    "@nestjs/common": "^10.2.7",
+    "@nestjs/core": "^10.2.7",
     "@nestjs/graphql": "^12.0.9",
-    "@nestjs/mongoose": "^9.2.2",
-    "@nestjs/platform-express": "^9.0.0",
+    "@nestjs/mongoose": "^10.0.1",
+    "@nestjs/platform-express": "^10.2.7",
     "graphql": "^16.8.1",
     "mongoose": "^7.0.3",
     "reflect-metadata": "^0.1.13",
-    "rxjs": "^7.2.0",
+    "rxjs": "^7.8.1",
     "webpack": "^5.0.0"
   },
   "devDependencies": {
-    "@nestjs/cli": "^9.0.0",
-    "@nestjs/schematics": "^9.0.0",
-    "@nestjs/testing": "^9.0.0",
+    "@nestjs/cli": "^10.1.18",
+    "@nestjs/schematics": "^10.0.2",
+    "@nestjs/testing": "^10.2.7",
     "@types/express": "^4.17.13",
-    "@types/jest": "29.2.4",
+    "@types/jest": "^29.5.6",
     "@types/node": "18.11.18",
     "@types/supertest": "^2.0.11",
     "eslint-config-custom": "*",
-    "jest": "29.3.1",
+    "jest": "^29.7.0",
     "source-map-support": "^0.5.20",
     "supertest": "^6.1.3",
-    "ts-jest": "29.0.3",
-    "ts-loader": "^9.2.3",
+    "ts-jest": "^29.1.1",
     "ts-node": "^10.0.0",
     "tsconfig": "*",
     "tsconfig-paths": "4.1.1",
     "typescript": "^5.1.3"
-  },
-  "jest": {
-    "moduleFileExtensions": [
-      "js",
-      "json",
-      "ts"
-    ],
-    "rootDir": "src",
-    "testRegex": ".*\\.spec\\.ts$",
-    "transform": {
-      "^.+\\.(t|j)s$": "ts-jest"
-    },
-    "collectCoverageFrom": [
-      "**/*.(t|j)s"
-    ],
-    "coverageDirectory": "../coverage",
-    "testEnvironment": "node"
   }
 }
