+++ apps/backend/src/cards/cards.resolver.ts
@@ -1,9 +1,9 @@
 import { Args, Query, Resolver } from '@nestjs/graphql';
-import { Card } from 'src/shared/schemas/card.schema';
 
 import { CardRefsByAge } from './dto/card-refs-by-age.dto';
 import { CardIdAndRefByAge } from './dto/cardId-and-ref-by-age.dto';
 import { FindOneOptionsInput } from './dto/find-one-options.input';
+import { Card } from './schemas/card.schema';
 import { CardsSortingService } from './services/cards-sorting.service';
 import { CardsService } from './services/cards.service';
