+++ apps/backend/src/cards/dto/create-card.dto.ts
@@ -1,5 +1,5 @@
-import { DogmaEffect, ResourceSpaces } from 'src/shared/schemas/card.schema';
-import { ResourceTotals } from 'src/shared/schemas/resource-totals.schema';
+import { DogmaEffect, ResourceSpaces } from '../schemas/card.schema';
+import { ResourceTotals } from '../schemas/resource-totals.schema';
 
 export class CreateCardDto {
   readonly cardId!: string;
