+++ apps/backend/src/cards/services/cards-sorting.service.ts
@@ -1,9 +1,10 @@
 import { Injectable } from '@nestjs/common';
-import { AgeString, cardAgeToAgeStringMap } from 'src/shared/constants/cards';
-import { Card } from 'src/shared/schemas/card.schema';
+
+import { AgeString, cardAgeToAgeStringMap } from '@inno/constants';
 
 import { CardRefsByAge } from '../dto/card-refs-by-age.dto';
 import { CardIdAndRefByAge } from '../dto/cardId-and-ref-by-age.dto';
+import { Card } from '../schemas/card.schema';
 
 @Injectable()
 export class CardsSortingService {
