+++ apps/backend/src/cards/services/cards.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
-import { Card, CardDocument } from 'src/shared/schemas/card.schema';
 
 import { CreateCardDto } from '../dto/create-card.dto';
+import { Card, CardDocument } from '../schemas/card.schema';
 
 @Injectable()
 export class CardsService {
