+++ apps/backend/src/player-game-details/schemas/player-game-details.schema.ts
@@ -1,7 +1,7 @@
 import { Field, ID, InputType, ObjectType } from '@nestjs/graphql';
 import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
 import { HydratedDocument, Schema as MongooseSchema } from 'mongoose';
-import { ResourceTotals } from 'src/shared/schemas/resource-totals.schema';
+import { ResourceTotals } from 'src/cards/schemas/resource-totals.schema';
 
 import { Board } from './board.schema';
